001 | add UUID

+++ src/main/java/ishmitko/app/tamagochibackend/repositories/AnimalRepository.java
@@ -3,5 +3,5 @@ package ishmitko.app.tamagochibackend.repositories;
 import ishmitko.app.tamagochibackend.model.Animal;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AnimalRepository extends JpaRepository<Animal, Long> {
+public interface AnimalRepository extends JpaRepository<Animal, Integer> {
 }
